--- js/data/comments.js
@@ -1,21 +1,0 @@
-import { getRandomInt, getRandomArrayElement } from './../util.js';
-import { NAMES } from './const.js';
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают.',
-  'Как можно было поймать такой неудачный момент?!',
-];
-
-const createComments = (_, index) => ({
-  name: getRandomArrayElement(NAMES),
-  id: index + 1,
-  avatar: `img/avatar-${getRandomInt(1, 6)}.svg`,
-  message: getRandomArrayElement(MESSAGES),
-});
-
-export const similarComments = () => Array.from({ length: getRandomInt(1, 3) }, createComments);
