--- js/data/const.js
@@ -1,3 +1,0 @@
-export const NAMES = [
-  'Вася', 'Маша', 'Даша', 'Глаша', 'Саша', 'Мася', 'Дуся', 'Муся',
-];
